Coin Sorter and Wrapper
Added code that uses nested for loops that utilizes a series of parallel arrays to calculate a running total dollar amount of all the coins combined.

The program calculate how many coin wrappers are required to wrap and store the coins based on when the coin count exceeds the coin wrapper's capacity.