Code Efficiency Update
Removed an older, more inefficient method of calculations for the program in favor of a quicker, neater method.

Amounts for each coin are now tracked in a parallel array like the program's other counterparts.

Removed the "coinCount" variable. The program now uses coinsAmount[i] in its stead. This will be reverted or revised if the program needs to keep a persistent track of the amounts for each coin type.